dividend added

+++ migrations/2020-03-21-130459_create_dividend/up.sql
@@ -4,6 +4,6 @@ CREATE TABLE dividend (
   payment_date DATE NOT NULL,
   announcement_date DATE NOT NULL,
   exdividend_date DATE NOT NULL,
-  payment MONEY NOT NULL,
+  payment FLOAT NOT NULL,
   PRIMARY KEY (company, payment_date)
 )
